CDP-493: Check md5 (if present) before downloading in case a download is not necessary
Restructured the inside of the transferAsset function so that logic is contained in the Promise in order to best utiltize resolve and reject.
Added logic that checks the md5 (if available) on an asset to see if an update is required before moving on to downloading the asset to do the same check again (if needed).