CDP-522: Add language specific categories for video
Added additional routes to Taxonomy (mainly for testing): find by name, and translate bv id and locale.
Added translateCategory middleware that asks the model for it's "units" and translates the categories based on the unit's language locale. Essentially a unit is just something with a locale that needs to have cateogory names translated and inserted into the object. The put method then puts the unit back in it's place via the model. The root categories property is then removed.