CDP-552: Create bulk taxonomy term import endpoint
Updated findTermByName to require a locale (or default to 'en'). This allows for an exact match on the term name in a specific language.
Added a sort on constructTree so that the tree is alphabetical.
Modified getAllDocuments to better handle chunking (since there is a hard limit of 10000 docs per result set).
Added back the primary reliance on a 'seen' array when importing terms in the event that ES hasn't propagated previous term indexings in the same process.
Added comments.