HOT FIX: Error handling when temp files get prematurely deleted.
For AWS: added on error handler for when the read stream gets interupted because the file got deleted. Upload aborts if so.
For CF: wrapped readstream in a passthrough as tus likes it better and may allow us to catch file read errors. Needs to be tested further.
Improved error handling elsewhere too (download for instance).